+++ modules/@apostrophecms/doc/ui/apos/apps/AposDoc.js
@@ -10,13 +10,18 @@ export default () => {
   // `_id` should be the `_id` of the existing document to edit; leave
   // blank to create a new document.
   //
+  // `copyOfId` is an optional `_id` of an existing document from which properties
+  // should be copied.
+  //
   // `copyOf` is an optional, existing document from which properties should be copied.
+  // It is present for BC.
   //
   // On success, returns the new or updated document. If the modal is cancelled,
   // `undefined` is returned. Be sure to `await` the result.
   apos.doc.edit = async ({
     type,
     _id,
+    copyOfId,
     copyOf
   }) => {
     if (!type) {
@@ -29,10 +34,14 @@ export default () => {
     if (!modal) {
       throw new Error(`${type} is not a valid piece or page type, or cannot be edited`);
     }
+    copyOfId = copyOfId ?? copyOf?._id;
+    if (copyOf && !copyOfId) {
+      throw new Error('copyOf (deprecated) must be an object with a `_id` property, if possible pass `copyOfId` instead');
+    }
     return apos.modal.execute(modal, {
       moduleName: type,
       docId: _id,
-      copyOf
+      copyOfId
     });
   };
   // If you don't care about the returned value, you can emit an